+++ packages/manager/src/factories/volume.ts
@@ -12,6 +12,7 @@ export const volumeFactory = Factory.Sync.makeFactory<Volume>({
   updated: '2019-01-01',
   filesystem_path: '/mnt',
   linode_id: null,
+  linode_label: null,
   hardware_type: 'hdd',
 });