+++ packages/manager/src/features/Volumes/VolumeDrawer/AttachVolumeToLinodeForm.tsx
@@ -1,16 +1,10 @@
 import { Formik } from 'formik';
 import { Grant } from '@linode/api-v4/lib/account';
-import { attachVolume } from '@linode/api-v4/lib/volumes';
-import { pathOr } from 'ramda';
 import * as React from 'react';
 import { connect, MapDispatchToProps } from 'react-redux';
 import { compose } from 'recompose';
 import Form from 'src/components/core/Form';
-import withVolumesRequests, {
-  VolumesRequests,
-} from 'src/containers/volumesRequests.container';
 import { resetEventsPolling } from 'src/eventsPolling';
-import { MapState } from 'src/store/types';
 import { openForCreating } from 'src/store/volumeForm';
 import { getErrorMap } from 'src/utilities/errorUtils';
 import {
@@ -20,11 +14,13 @@ import {
 import { number, object } from 'yup';
 import ConfigSelect from './ConfigSelect';
 import { modes } from './modes';
-import ModeSelection from './ModeSelection';
+import { ModeSelection } from './ModeSelection';
 import NoticePanel from './NoticePanel';
 import Notice from 'src/components/Notice';
 import VolumesActionsPanel from './VolumesActionsPanel';
 import VolumeSelect from './VolumeSelect';
+import { useAttachVolumeMutation } from 'src/queries/volumes';
+import { useGrants } from 'src/queries/profile';
 
 interface Props {
   onClose: () => void;
@@ -34,38 +30,45 @@ interface Props {
   readOnly?: boolean;
 }
 
-type CombinedProps = Props & StateProps & DispatchProps & VolumesRequests;
+type CombinedProps = Props & DispatchProps;
 
 /**
  * I had to provide a separate validation schema since the linode_id (which is required by API) is
  * provided as a prop and not a user input value.
  */
-const validationScheme = object({
-  volume_id: number().required(),
-  config_id: number().required(),
+const AttachVolumeValidationSchema = object({
+  volume_id: number()
+    .min(0, 'Volume is required.')
+    .required('Volume is required.'),
+  config_id: number()
+    .min(0, 'Config is required.')
+    .required('Config is required.'),
 });
 
 const initialValues = { volume_id: -1, config_id: -1 };
 
 const AttachVolumeToLinodeForm: React.FC<CombinedProps> = (props) => {
-  const {
-    actions,
-    onClose,
-    linodeId,
-    linodeRegion,
-    linodeGrants,
-    readOnly,
-  } = props;
+  const { actions, onClose, linodeId, linodeRegion, readOnly } = props;
+
+  const { data: grants } = useGrants();
+
+  const linodeGrants = grants?.linode ?? [];
+
   const linodeGrant = linodeGrants.filter(
     (grant: Grant) => grant.id === linodeId
   )[0];
+
   const disabled =
     readOnly || (linodeGrant && linodeGrant.permissions !== 'read_write');
+
+  const { mutateAsync: attachVolume } = useAttachVolumeMutation();
+
   return (
     <Formik
-      validationSchema={validationScheme}
+      validationSchema={AttachVolumeValidationSchema}
       onSubmit={(values, { setSubmitting, setStatus, setErrors }) => {
-        attachVolume(values.volume_id, {
+        attachVolume({
+          volumeId: values.volume_id,
           linode_id: linodeId,
           config_id: values.config_id,
         })
@@ -183,20 +186,8 @@ const mapDispatchToProps: MapDispatchToProps<DispatchProps, Props> = (
   },
 });
 
-interface StateProps {
-  linodeGrants: Grant[];
-}
+const connected = connect(undefined, mapDispatchToProps);
 
-const mapStateToProps: MapState<StateProps, CombinedProps> = (state) => ({
-  linodeGrants: pathOr(
-    [],
-    ['__resources', 'profile', 'data', 'grants', 'linode'],
-    state
-  ),
-});
-
-const connected = connect(mapStateToProps, mapDispatchToProps);
-
-const enhanced = compose<CombinedProps, Props>(connected, withVolumesRequests);
+const enhanced = compose<CombinedProps, Props>(connected);
 
 export default enhanced(AttachVolumeToLinodeForm);