+++ packages/manager/src/features/Volumes/VolumeDrawer/VolumeDrawer.tsx
@@ -13,11 +13,11 @@ import {
   viewResizeInstructions,
 } from 'src/store/volumeForm';
 import AttachVolumeToLinodeForm from './AttachVolumeToLinodeForm';
-import CloneVolumeForm from './CloneVolumeForm';
+import { CloneVolumeForm } from './CloneVolumeForm';
 import CreateVolumeForLinodeForm from './CreateVolumeForLinodeForm';
-import EditVolumeForm from './EditVolumeForm';
+import { EditVolumeForm } from './EditVolumeForm';
 import { modes } from './modes';
-import ResizeVolumeForm from './ResizeVolumeForm';
+import { ResizeVolumeForm } from './ResizeVolumeForm';
 import ResizeVolumesInstruction from './ResizeVolumesInstruction';
 import VolumeConfigForm from './VolumeConfigForm';