+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2022-11-30] - v1.81.0
+
+### Added:
+- SMTP restriction notice helper text
+- Object Storage type to confirm
+- Copyable Linode IP Address text
+- Copyable Kubernetes Node Pool IP address text
+- React Query for Volumes
+
+### Fixed:
+- Landing page empty state 2-column layout
+
 ## [2022-11-14] - v1.80.0
 
 ### Added: